peerdefine/requirements.txt
addict==2.1.3
amqp==2.3.2
argon2-cffi==18.1.0
bcrypt==3.1.4
bibtexparser==1.0.1
billiard==3.5.0.4
celery==4.2.1
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
Click==7.0
defusedxml==0.5.0
Django==2.0.8
django-allauth==0.37.1
django-appconf==1.0.2
-e git+https://github.com/yourlabs/django-autocomplete-light.git@13529dd158769c3ea29b6c7bade9d06cb70acf34#egg=django_autocomplete_light
django-bootstrap3==9.1.0
django-bower==5.2.0
django-braces==1.12.0
django-ckeditor==5.6.1
django-crispy-forms==1.7.2
django-debug-toolbar==1.9.1
django-extensions==2.1.2
django-extra-views==0.10.0
django-floppyforms==1.7.1.dev1
django-grappelli==2.11.1
django-js-asset==1.1.0
django-model-utils==3.1.2
django-nested-admin==3.0.21
django-notifications-hq==1.5.0
django-reversion==2.0.13
django-select2==6.0.3
django-select2-forms==2.0.2
django-sortedm2m==1.5.0
django-tagulous==0.13.2
dparse==0.4.1
Faker==0.8.12
future==0.16.0
google-scholar-scraper==0.2
habanero==0.6.2
idna==2.6
jsonfield==2.0.2
kombu==4.2.1
misaka==2.1.0
nodeenv==1.3.0
oauthlib==2.1.0
packaging==18.0
Pillow==5.0.0
pycparser==2.18
pyparsing==2.2.0
python-dateutil==2.7.0
python-monkey-business==1.0.0
python3-openid==3.1.0
pytz==2018.3
PyYAML==3.13
requests==2.20.0
requests-oauthlib==1.0.0
safety==1.8.4
six==1.11.0
sqlparse==0.2.4
text-unidecode==1.2
urllib3==1.22
vine==1.1.4
